// src/platform/packages/shared/kbn-lens-embeddable-utils/config_builder/types.d.ts
export type DataViewsCommon = Pick<DataViewsService, 'get' | 'create'>;
export type LensAttributes = TypedLensByValueInput['attributes'];
export type DeepWriteable<T> = {
    -readonly [P in keyof T]: DeepWriteable<T[P]>;
};
type Identity<T> = T extends object ? {
    [P in keyof T]: T[P];
} : T;
export type ChartType = 'xy' | 'pie' | 'heatmap' | 'metric' | 'gauge' | 'donut' | 'mosaic' | 'regionmap' | 'table' | 'tagcloud' | 'treemap';
export interface TimeRange {
    from: string;
    to: string;
    type: 'relative' | 'absolute';
}
export type LensLayerQuery = string;
export interface LensDataviewDataset {
    index: string;
    timeFieldName?: string;
}
export type LensDatatableDataset = Datatable;
export interface LensESQLDataset {
    esql: string;
}
export type LensDataset = LensDataviewDataset | LensDatatableDataset | LensESQLDataset;
export interface LensBaseConfig {
    title: string;
    description?: string;
    /** default data view id or index pattern to use, it can be overriden on each query */
    dataset?: LensDataset;
}
export interface LensBaseLayer {
    label?: string;
    filter?: string;
    format?: 'bits' | 'bytes' | 'currency' | 'duration' | 'number' | 'percent' | 'string';
    decimals?: number;
    normalizeByUnit?: 's' | 'm' | 'h' | 'd';
    suffix?: string;
    fromUnit?: string;
    toUnit?: string;
    compactValues?: boolean;
    randomSampling?: number;
    useGlobalFilter?: boolean;
    seriesColor?: string;
    value: LensLayerQuery;
}
// ...
export interface LensLegendConfig {
    show?: boolean;
    position?: 'top' | 'left' | 'bottom' | 'right';
    legendStats?: XYLegendValue[];
}
// ...
export type LensBreakdownConfig = string | Identity<(LensBreakdownTopValuesConfig | LensBreakdownIntervalsConfig | LensBreakdownFiltersConfig | LensBreakdownDateHistogramConfig) & {
    colorPalette?: string;
}>;
// ...
export interface LensHeatmapConfigBase {
    chartType: 'heatmap';
    /** field name to apply breakdown based on field type or full breakdown configuration */
    breakdown?: LensBreakdownConfig;
    xAxis: LensBreakdownConfig;
    legend?: Identity<LensLegendConfig>;
}
export type LensHeatmapConfig = Identity<LensBaseConfig & LensBaseLayer & LensHeatmapConfigBase>;
// ...
export interface BuildDependencies {
    dataViewsAPI: DataViewsCommon;
}